requirements.txt
pillow==11.2.1
google-genai>=1.19.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-google-genai==2.1.5
langchain-tavily==0.2.0
langchain-text-splitters==0.3.8
langfuse==2.60.7
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.43